fix: fix addBalasan dan deleteBalasan

+++ src/app/reply/treebalasan.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 #include <stdlib.h>
-#include "treeBalasan.h"
+#include "treebalasan.h"
 
 AddressTree createNewNode(Balasan balasan)
 {
-    NodeBalasan *newBalasan = (NodeBalasan *)malloc(sizeof(NodeBalasan));
+    AddressTree newBalasan = (AddressTree) malloc(sizeof(NodeBalasan));
     if (newBalasan != NULL)
     {
         INFO(newBalasan) = balasan;
@@ -21,7 +21,7 @@ void createTreeBalasan(TreeBalasan *t)
 
 void addBalasan(AddressTree *n, Balasan balasan)
 {
-    NodeBalasan *newBalasan = createNewNode(balasan);
+    AddressTree newBalasan = createNewNode(balasan);
     if (isTreeBalasanEmpty(*n))
     {
         // printf("not ok 1\n");
@@ -35,8 +35,8 @@ void addBalasan(AddressTree *n, Balasan balasan)
     }
     else
     {
-        printf("od\n");
-        AddressTree p = (*n);
+        // printf("od\n");
+        AddressTree p = LEFT(*n);
         while (RIGHT(p) != NULL)
         {
             p = RIGHT(p);
@@ -46,62 +46,101 @@ void addBalasan(AddressTree *n, Balasan balasan)
     }
 }
 
-void deleteTreeBalasan(TreeBalasan *t, int idFind)
-{
-    AddressTree p;
-    boolean isFound = false;
-    if (searchBalasan(*t, idFind) != NULL)
-    {
-        printf("Info hasil search: %d\n", IDBALASAN(INFO(searchBalasan(*t, idFind))));
-        if (IDBALASAN(INFO(*t)) == idFind)
-        {
-            printf("Masuk diri sendiri\n");
-            deleteAllNode(LEFT(*t));
-            printf("Info left sekarang: %d\n", IDBALASAN(INFO(RIGHT(*t))));
-            p = *t;
+// void deleteTreeBalasan(TreeBalasan *t, int idFind)
+// {
+//     AddressTree p;
+//     boolean isFound = false;
+//     if (searchBalasan(*t, idFind) != NULL)
+//     {
+//         printf("Info hasil search: %d\n", IDBALASAN(INFO(searchBalasan(*t, idFind))));
+//         if (IDBALASAN(INFO(*t)) == idFind)
+//         {
+//             printf("Masuk diri sendiri\n");
+//             deleteAllNode(LEFT(*t));
+//             printf("Info left sekarang: %d\n", IDBALASAN(INFO(RIGHT(*t))));
+//             p = *t;
 
-            LEFT(*t) = RIGHT(*t);
+//             LEFT(*t) = RIGHT(*t);
 
-            RIGHT(*t) = RIGHT(LEFT(*t));
+//             RIGHT(*t) = RIGHT(LEFT(*t));
 
-            free(p);
+//             free(p);
+//         }
+//         else if (searchBalasan(RIGHT(*t), idFind) != NULL)
+//         {
+//             deleteTreeBalasan(&RIGHT(*t), idFind);
+//             //     else // Kalo di RIGHT(t) gaada, pasti RIGHT(t) punya RIGHT lagi
+//             //     {
+//             //         AddressTree prevDel = RIGHT(t);
+//             //         AddressTree del = RIGHT(prevDel);
+//             //         while (!isFound && del != NULL)
+//             //         {
+//             //             if (IDBALASAN(INFO(del)) == idFind)
+//             //             {
+//             //                 RIGHT(prevDel) = RIGHT(del);
+//             //                 deleteAllNode(del);
+//             //                 isFound = true;
+//             //             }
+//             //             else if (searchBalasan(del, idFind) != 0)
+//             //             {
+//             //                 deleteTreeBalasan(del, idFind);
+//             //                 isFound = true;
+//             //             }
+//             //             prevDel = RIGHT(prevDel);
+//             //             del = RIGHT(del);
+//             //         }
+//             //     }
+//         }
+//         else
+//         { // Balasan yang ingin dihapus ada di LEFT(t)
+//             printf("Masuk kiri bang\n");
+//             deleteTreeBalasan(&LEFT(*t), idFind);
+//         }
+//     }
+// }
+
+boolean deleteTreeBalasan(TreeBalasan t, int idFind){
+    AddressTree p = t;
+    boolean udah = false;
+    if (IDBALASAN(INFO(t)) == idFind){
+        t = RIGHT(p);
+        deleteAllNode(LEFT(p));
+        free(p);
+        return true;
+    }
+    else{
+        AddressTree sibling = RIGHT(p);
+        if (sibling != NULL){
+            if (IDBALASAN(INFO(sibling)) == idFind){
+                RIGHT(p) = RIGHT(sibling);
+                deleteAllNode(LEFT(sibling));
+                free(sibling);
+                return true;
+            }
+            else{
+                // printf("masuk sini\n");
+                udah = deleteTreeBalasan(sibling, idFind);
+            }
         }
-        else if (searchBalasan(RIGHT(*t), idFind) != NULL)
-        {
-            deleteTreeBalasan(&RIGHT(*t), idFind);
-            //     else // Kalo di RIGHT(t) gaada, pasti RIGHT(t) punya RIGHT lagi
-            //     {
-            //         AddressTree prevDel = RIGHT(t);
-            //         AddressTree del = RIGHT(prevDel);
-            //         while (!isFound && del != NULL)
-            //         {
-            //             if (IDBALASAN(INFO(del)) == idFind)
-            //             {
-            //                 RIGHT(prevDel) = RIGHT(del);
-            //                 deleteAllNode(del);
-            //                 isFound = true;
-            //             }
-            //             else if (searchBalasan(del, idFind) != 0)
-            //             {
-            //                 deleteTreeBalasan(del, idFind);
-            //                 isFound = true;
-            //             }
-            //             prevDel = RIGHT(prevDel);
-            //             del = RIGHT(del);
-            //         }
-            //     }
-        }
-        else
-        { // Balasan yang ingin dihapus ada di LEFT(t)
-            printf("Masuk kiri bang\n");
-            deleteTreeBalasan(&LEFT(*t), idFind);
+        AddressTree child = LEFT(p);
+        if (child != NULL && !udah){
+            if (IDBALASAN(INFO(child)) == idFind){
+                LEFT(p) = RIGHT(child);
+                deleteAllNode(LEFT(child));
+                free(child);
+                return true;
+            }
+            else{
+                udah = deleteTreeBalasan(sibling, idFind);
+            }
         }
     }
+    return udah;
 }
 
-void deleteAllNode(NodeBalasan *n)
+void deleteAllNode(AddressTree n)
 {
-    printf("line102\n");
+    // printf("line102\n");
     if (!isTreeBalasanEmpty(n))
     {
         if (isOneElementTree(n))
@@ -114,19 +153,26 @@ void deleteAllNode(NodeBalasan *n)
         else
         {
             printf("Masuk banyakelmt\n");
-            while (LEFT(n) != NULL)
-            {
+            // while (LEFT(n) != NULL)
+            // {
+            //     deleteAllNode(LEFT(n));
+            // }
+            // while (RIGHT(n) != NULL)
+            // {
+            //     deleteAllNode(RIGHT(n));
+            // }
+            if (LEFT(n) != NULL){
                 deleteAllNode(LEFT(n));
             }
-            while (RIGHT(n) != NULL)
-            {
+            if (RIGHT(n) != NULL){
                 deleteAllNode(RIGHT(n));
             }
             // free(n);
         }
     }
-    free(n);
-    printf("line124\n");
+    if (n != NULL)
+        free(n);
+    // printf("line124\n");
 }
 
 boolean isTreeBalasanEmpty(TreeBalasan t)
